BACKLOG-12335 : Reorganized reducers
