+++ src/javascript/JContent/ContentRoute/ContentLayout/ContentListTable/ContentListTable.jsx
@@ -12,17 +12,17 @@ import * as _ from 'lodash';
 import {useTranslation} from 'react-i18next';
 import PublicationStatus from '../PublicationStatus';
 import dayjs from 'dayjs';
-import {CM_DRAWER_STATES, cmGoto, cmOpenPaths, cmSetMode} from '../../../JContent.redux-actions';
+import {CM_DRAWER_STATES, cmGoto, cmOpenPaths, cmSetMode} from '../../../JContent.redux';
 import {allowDoubleClickNavigation, extractPaths, getDefaultLocale, isMarkedForDeletion} from '../../../JContent.utils';
 import ToolBar from '../ToolBar';
 import {connect} from 'react-redux';
 import {compose} from 'react-apollo';
 import UploadTransformComponent from '../UploadTransformComponent';
 import classNames from 'classnames';
-import {cmSetPreviewSelection} from '../../../preview.redux-actions';
-import {cmSetSort} from '../sort.redux-actions';
-import {cmSetPage, cmSetPageSize} from '../pagination.redux-actions';
-import {cmAddSelection, cmRemoveSelection, cmSwitchSelection} from '../contentSelection.redux-actions';
+import {cmSetPreviewSelection} from '../../../preview.redux';
+import {cmSetSort} from '../sort.redux';
+import {cmSetPage, cmSetPageSize} from '../pagination.redux';
+import {cmAddSelection, cmRemoveSelection, cmSwitchSelection} from '../contentSelection.redux';
 import JContentConstants from '../../../JContent.constants';
 import ContentListEmptyDropZone from './ContentListEmptyDropZone';
 import ContentNotFound from './ContentNotFound';
