+++ src/javascript/JContent/ContentRoute/ContentLayout/FilesGrid/FilesGrid.jsx
@@ -9,10 +9,10 @@ import {Typography} from '@jahia/design-system-kit';
 import {Pagination, DxContext} from '@jahia/react-material';
 import UploadTransformComponent from '../UploadTransformComponent';
 import {connect} from 'react-redux';
-import {cmSetPage, cmSetPageSize} from '../pagination.redux-actions';
+import {cmSetPage, cmSetPageSize} from '../pagination.redux';
 import FilesGridEmptyDropZone from './FilesGridEmptyDropZone';
-import {cmSetPreviewSelection} from '../../../preview.redux-actions';
-import {cmGoto, cmOpenPaths} from '../../../JContent.redux-actions';
+import {cmSetPreviewSelection} from '../../../preview.redux';
+import {cmGoto, cmOpenPaths} from '../../../JContent.redux';
 import classNames from 'classnames';
 import {extractPaths} from '../../../JContent.utils';
 import {useKeyboardNavigation} from '../useKeyboardNavigation';
