--- src/javascript/JContent/ContentRoute/ContentLayout/FilesGrid/FilesGrid.redux-actions.js
@@ -1,4 +1,0 @@
-import {createActions} from 'redux-actions';
-
-export const {filesgridSetMode, filesgridSetGridMode} = createActions('FILESGRID_SET_MODE', 'FILESGRID_SET_GRID_MODE');
-/**/
