+++ src/javascript/JContent/ContentRoute/ContentLayout/SearchBar/SearchBarNormal/SearchBarNormal.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import {Input, withStyles} from '@material-ui/core';
 import ContentTypeSelect from './ContentTypeSelect';
-import {cmGoto} from '../../../../JContent.redux-actions';
+import {cmGoto} from '../../../../JContent.redux';
 import {withNotifications, DxContext} from '@jahia/react-material';
 import {withTranslation} from 'react-i18next';
 import {compose} from 'react-apollo';
