--- src/javascript/JContent/ContentRoute/ContentLayout/pagination.redux-actions.js
@@ -1,3 +1,0 @@
-import {createActions} from 'redux-actions';
-
-export const {cmSetPage, cmSetPageSize} = createActions('CM_SET_PAGE', 'CM_SET_PAGE_SIZE');
