--- src/javascript/JContent/ContentRoute/ContentLayout/sort.redux-actions.js
@@ -1,3 +1,0 @@
-import {createAction} from 'redux-actions';
-
-export const cmSetSort = createAction('CM_SET_SORT');
