--- src/javascript/JContent/ContentRoute/ContentLayout/sort.redux-reducers.js
@@ -1,4 +1,0 @@
-import {cmSetSort} from './sort.redux-actions';
-import {handleAction} from 'redux-actions';
-
-export const sortReducer = handleAction(cmSetSort, (state, action) => (action.payload), {order: 'ASC', orderBy: 'lastModified.value'});
