--- src/javascript/JContent/JContent.redux-utils.js
@@ -1,127 +1,0 @@
-import {cmGoto} from './JContent.redux-actions';
-import _ from 'lodash';
-import rison from 'rison';
-import queryString from 'query-string';
-import {push} from 'connected-react-router';
-import JContentConstants from './JContent.constants';
-
-const PARAMS_KEY = '?params=';
-const DEFAULT_MODE_PATHS = {browse: '/contents', media: '/files'};
-let currentValue;
-
-let select = state => {
-    let {router: {location: {pathname, search}}, site, language, mode, path, params} = state;
-    return {
-        pathname,
-        search,
-        site,
-        language,
-        mode,
-        path,
-        params
-    };
-};
-
-let buildUrl = (site, language, mode, path, params) => {
-    let sitePath = '/sites/' + site;
-    if (path.startsWith(sitePath + '/')) {
-        path = path.substring(('/sites/' + site).length);
-    } else if (mode === JContentConstants.mode.APPS) {
-        // Path is an action key
-        path = path.startsWith('/') ? path : '/' + path;
-    } else {
-        path = '';
-    }
-
-    // Special chars in folder naming
-    path = path.replace(/[^/]/g, encodeURIComponent);
-
-    let queryString = _.isEmpty(params) ? '' : PARAMS_KEY + rison.encode_uri(params);
-    return '/jcontent/' + [site, language, mode].join('/') + path + queryString;
-};
-
-let extractParamsFromUrl = (pathname, search) => {
-    if (pathname.startsWith('/jcontent')) {
-        let [, , site, language, mode, ...pathElements] = pathname.split('/');
-
-        let path;
-        if (mode === JContentConstants.mode.APPS) {
-            path = pathElements.join('/');
-        } else {
-            path = '/sites/' + site;
-            if (_.isEmpty(pathElements)) {
-                if (mode === JContentConstants.mode.MEDIA) {
-                    path += '/files';
-                }
-            } else {
-                path += ('/' + pathElements.join('/'));
-            }
-        }
-
-        path = decodeURIComponent(path);
-        let params = deserializeQueryString(search);
-        return {site, language, mode, path, params};
-    }
-
-    return {site: '', language: '', mode: '', path: '', params: {}};
-};
-
-let deserializeQueryString = search => {
-    if (search) {
-        let params = queryString.parse(search).params;
-        if (params) {
-            try {
-                return rison.decode(params);
-                // eslint-disable-next-line no-unused-vars
-            } catch (e) {
-                return {};
-            }
-        }
-    }
-
-    return {};
-};
-
-let pathResolver = (currentValue, currentValueFromUrl) => {
-    if (currentValue.site !== currentValueFromUrl.site && currentValue.mode !== JContentConstants.mode.APPS) {
-        // Switched sites, we have to set default path based on mode: browse -> /content-folders | media -> /files
-        return currentValueFromUrl.path.substr(0, currentValueFromUrl.path.indexOf(currentValueFromUrl.site)) + currentValue.site + DEFAULT_MODE_PATHS[currentValue.mode];
-    }
-
-    return currentValue.path;
-};
-
-let getSyncListener = store => () => {
-    setTimeout(() => {
-        let previousValue = currentValue || {};
-        currentValue = select(store.getState());
-        let currentValueFromUrl = extractParamsFromUrl(currentValue.pathname, currentValue.search);
-        if (previousValue.pathname !== currentValue.pathname || previousValue.search !== currentValue.search) {
-            if (currentValueFromUrl.site !== previousValue.site ||
-                currentValueFromUrl.language !== previousValue.language ||
-                currentValueFromUrl.mode !== previousValue.mode ||
-                currentValueFromUrl.path !== previousValue.path ||
-                !_.isEqual(currentValueFromUrl.params, previousValue.params)
-            ) {
-                let data = {};
-                Object.assign(data,
-                    currentValueFromUrl.site === previousValue.site ? {} : {site: currentValueFromUrl.site},
-                    currentValueFromUrl.language === previousValue.language ? {} : {language: currentValueFromUrl.language},
-                    currentValueFromUrl.mode === previousValue.mode ? {} : {mode: currentValueFromUrl.mode},
-                    currentValueFromUrl.path === previousValue.path ? {} : {path: currentValueFromUrl.path},
-                    _.isEqual(currentValueFromUrl.params, previousValue.params) ? {} : {params: currentValueFromUrl.params}
-                );
-                store.dispatch(cmGoto(data));
-            }
-        } else if ((previousValue.site !== currentValue.site && currentValueFromUrl.site !== currentValue.site) ||
-            (previousValue.language !== currentValue.language && currentValueFromUrl.language !== currentValue.language) ||
-            (previousValue.mode !== currentValue.mode && currentValueFromUrl.mode !== currentValue.mode) ||
-            (previousValue.path !== currentValue.path && currentValueFromUrl.path !== currentValue.path) ||
-            (!_.isEqual(currentValueFromUrl.params, currentValue.params))
-        ) {
-            store.dispatch(push(buildUrl(currentValue.site, currentValue.language, currentValue.mode, encodeURI(pathResolver(currentValue, currentValueFromUrl)), currentValue.params)));
-        }
-    });
-};
-
-export {getSyncListener, extractParamsFromUrl};
