--- src/javascript/JContent/actions/copyPaste/copyPaste.redux-actions.js
@@ -1,3 +1,0 @@
-import {createActions} from 'redux-actions';
-
-export const {copypasteCopy, copypasteCut, copypasteClear} = createActions('COPYPASTE_COPY', 'COPYPASTE_CUT', 'COPYPASTE_CLEAR');
