+++ src/javascript/JContent/preview.redux.js
@@ -1,0 +1,30 @@
+import {CM_DRAWER_STATES, CM_PREVIEW_MODES} from './JContent.redux';
+import {createActions, handleAction, handleActions} from 'redux-actions';
+
+export const {cmSetPreviewMode, cmSetPreviewState} = createActions('CM_SET_PREVIEW_MODE', 'CM_SET_PREVIEW_STATE');
+
+export const cmSetPreviewSelection = previewSelection => (dispatch, getState) => {
+    if (!previewSelection || getState().selection.length === 0) {
+        dispatch({
+            type: 'CM_SET_PREVIEW_SELECTION',
+            payload: previewSelection
+        });
+    }
+};
+
+cmSetPreviewSelection.toString = () => 'CM_SET_PREVIEW_SELECTION';
+
+export const previewReduxReducers = registry => {
+    const previewSelectionReducer = handleAction(cmSetPreviewSelection, (state, action) => action.payload, null);
+
+    const previewModeReducer = handleActions({
+        [cmSetPreviewMode]: (state, action) => action.payload,
+        [cmSetPreviewState]: () => CM_PREVIEW_MODES.EDIT
+    }, CM_PREVIEW_MODES.EDIT);
+
+    const previewStateReducer = handleAction(cmSetPreviewState, (state, action) => action.payload, CM_DRAWER_STATES.HIDE);
+
+    registry.add('redux-reducer', 'previewMode', {reducer: previewModeReducer});
+    registry.add('redux-reducer', 'previewState', {reducer: previewStateReducer});
+    registry.add('redux-reducer', 'previewSelection', {reducer: previewSelectionReducer});
+};
