+++ src/javascript/register.jsx
@@ -7,9 +7,16 @@ import JContentApp from './JContentApp';
 import Collections from '@jahia/moonstone/dist/icons/Collections';
 import contentManagerRoutes from './JContent/JContent.routes';
 import jContentActions from './JContent/JContent.actions';
-import jContentReduxStore from './JContent/JContent.redux-store';
 import jContentAccordionItems from './JContent/JContent.accordion-items';
 import {jContentAppRoot} from './JContent/JContent.app-root';
+import {jContentReduxReducers} from './JContent/JContent.redux';
+import {fileuploadReduxReducers} from './JContent/ContentRoute/ContentLayout/Upload/Upload.redux';
+import {previewReduxReducers} from './JContent/preview.redux';
+import {copypasteReduxReducers} from './JContent/actions/copyPaste/copyPaste.redux';
+import {filesGridReduxReducers} from './JContent/ContentRoute/ContentLayout/FilesGrid/FilesGrid.redux';
+import {paginationReduxReducers} from './JContent/ContentRoute/ContentLayout/pagination.redux';
+import {sortReduxReducers} from './JContent/ContentRoute/ContentLayout/sort.redux';
+import {contentSelectionReduxReducers} from './JContent/ContentRoute/ContentLayout/contentSelection.redux';
 
 const ROUTE = '/jcontent';
 const SYSTEM_SITE_ROUTE = `${ROUTE}/${window.contextJsParameters.siteKey}/${window.contextJsParameters.locale}/browse`;
@@ -42,7 +49,16 @@ registry.add('callback', 'jContent', {
 
         contentManagerRoutes(registry);
         jContentActions(registry);
-        jContentReduxStore(registry, window.contextJsParameters);
+
+        jContentReduxReducers(registry);
+        fileuploadReduxReducers(registry);
+        previewReduxReducers(registry);
+        copypasteReduxReducers(registry);
+        filesGridReduxReducers(registry);
+        paginationReduxReducers(registry);
+        sortReduxReducers(registry);
+        contentSelectionReduxReducers(registry);
+
         jContentAccordionItems(registry);
         jContentAppRoot(registry);
     }
